Bug fixes for pagination and combined CorpusID & keyword search
Fixed a bug for pagination in which updates to the search query did not result in immediate change to the cards displayed.

Fixed a bug in which searching for a new query while on a non-first page of an existing query did not reset the current page to 1.

Removed option for user to switch between querying keywords and CorpusID. Now all numerical inputs will be treated as keywords and everything else as a keyword query.