Bug fix for displaying CorpusID search
Fixed bug in which searching between different CorpusIDs would not update the card components. This is because the passed in prop was not updated.

Bug fix for displaying CorpusID search

Fixed bug in which searching between different CorpusIDs would not update the card components. This is because the passed in prop was not updated.